feat: extract department logic into a custom hook and separate logic from UI
